Made the Regular concept checking complete.
As the Regular concept is the foundation of generic programming with value
semantics, it is necessary to be able to automatically check if a type
respect these constraints.
It is now checked if == is an equivalence, and < a total ordering.
Predicates have been added to check these properties on a relation.
isRegular now take as input a range of values, which ideally contains all
the values in the type to be tested for regularity.
added.

Change-Id: Id2e4935ef22bce890796d92eb0c738e641d99b36
Reviewed-on: http://gerrit.aldebaran.lan/73127
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Tested-by: jmonnon <jmonnon@aldebaran.com>
Reviewed-by: jlamotte <jlamotte@aldebaran.com>