Do not throw in FutureSync destructor anymore
* FutureSync must synchronize on destruction to ensure that the task is finished.
Until now, it did a future.value() which caused an exception to be thrown in case of error.
Now, a future.wait() is performed, meaning any error will be "swallowed".
This prevents that 2 exceptions run at once (and the abort that follows).
If the user wants to check for error, he can call explicitly wait() before destruction,
or assign to the underlying type.

Change-Id: Icb5ef7d530b65da8e0649500501cd661590f3819
Reviewed-on: http://gerrit.aldebaran.lan/75114
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Tested-by: hcuche <hcuche@aldebaran.com>