Fix deadlock when disconnecting a signal from the callback
The deadlock would happen when disconnecting from the callback, if the
SignalSubscriber already had one active thread. The thread disconnecting
from the callback would wait for the inactive promise to be set.
This should normally happen when the first active thread terminates.

However, when it terminated, we checked if it was the last active thread
(and set the promise in this case only). That was incorrect, since there
was another activeThread (the one waiting for inactive).
The correct thing to check when removing an active thread is:
if all remaining threads are waiting for inactive, set inactive.

Change-Id: I8893e5ca8ae699028ce7d6b4cbe828a30305d687
Reviewed-on: http://gerrit.aldebaran.lan/77220
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Tested-by: jlamotte <jlamotte@aldebaran.com>