Fixed: copied an object containing qi::Atomics which must be shared.
This line triggers a rare race which involve the copied atomic counters
never reaching the value waited for, ending up in a either a dead or
live-lock, or a long execution.

This fix makes the test pass on Windows in debug version while before
it was often over the timeout.

Change-Id: I0b8fdba296877d52e38e032f222e1a5f309503c9
Reviewed-on: http://gerrit.aldebaran.lan/77732
Reviewed-by: tveltrop <tveltrop@aldebaran-robotics.com>
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Tested-by: jlamotte <jlamotte@aldebaran.com>