Removed warnings when compiling libqi.
`NOMINMAX` is already defined by CMake script for libqi and Visual Studio
generates a warning when a macro is defined several times for the
same compilation unit.
This patch make sure we define it only if it's not already defined.
It also removes intrinsic functions declarations that are already
defined in intrin.h.

Change-Id: I67baeb320e9d6dfb579c951f49b51b779f3339ae
Reviewed-on: http://gerrit.aldebaran.lan/80303
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Reviewed-by: sbarthelemy <sbarthelemy@aldebaran.com>
Tested-by: jlamotte <jlamotte@aldebaran.com>
Reviewed-by: jlamotte <jlamotte@aldebaran.com>